Move fee and fine related repositories to a separate package CIRC-849
