+++ src/main/java/org/folio/circulation/domain/anonymization/service/DefaultLoansFinder.java
@@ -5,7 +5,7 @@ import static java.util.concurrent.CompletableFuture.completedFuture;
 import java.util.Collection;
 import java.util.concurrent.CompletableFuture;
 
-import org.folio.circulation.infrastructure.storage.AccountRepository;
+import org.folio.circulation.infrastructure.storage.feesandfines.AccountRepository;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.MultipleRecords;
 import org.folio.circulation.support.Clients;
