+++ src/main/java/org/folio/circulation/infrastructure/storage/feesandfines/FeeFineActionRepository.java
@@ -1,4 +1,4 @@
-package org.folio.circulation.infrastructure.storage;
+package org.folio.circulation.infrastructure.storage.feesandfines;
 
 import static java.util.Objects.isNull;
 import static java.util.concurrent.CompletableFuture.allOf;
