+++ src/main/java/org/folio/circulation/services/LostItemFeeChargingService.java
@@ -18,8 +18,8 @@ import java.util.function.Function;
 import org.folio.circulation.StoreLoanAndItem;
 import org.folio.circulation.domain.FeeFine;
 import org.folio.circulation.domain.FeeFineOwner;
-import org.folio.circulation.infrastructure.storage.FeeFineOwnerRepository;
-import org.folio.circulation.infrastructure.storage.FeeFineRepository;
+import org.folio.circulation.infrastructure.storage.feesandfines.FeeFineOwnerRepository;
+import org.folio.circulation.infrastructure.storage.feesandfines.FeeFineRepository;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.infrastructure.storage.LostItemPolicyRepository;
 import org.folio.circulation.domain.policy.lostitem.LostItemPolicy;
