+++ src/test/java/org/folio/circulation/domain/OverdueFineCalculatorServiceTest.java
@@ -29,10 +29,10 @@ import org.folio.circulation.infrastructure.storage.OverdueFinePolicyRepository;
 import org.folio.circulation.domain.representations.CheckInByBarcodeRequest;
 import org.folio.circulation.domain.representations.StoredAccount;
 import org.folio.circulation.domain.representations.StoredFeeFineAction;
-import org.folio.circulation.infrastructure.storage.AccountRepository;
-import org.folio.circulation.infrastructure.storage.FeeFineActionRepository;
-import org.folio.circulation.infrastructure.storage.FeeFineOwnerRepository;
-import org.folio.circulation.infrastructure.storage.FeeFineRepository;
+import org.folio.circulation.infrastructure.storage.feesandfines.AccountRepository;
+import org.folio.circulation.infrastructure.storage.feesandfines.FeeFineActionRepository;
+import org.folio.circulation.infrastructure.storage.feesandfines.FeeFineOwnerRepository;
+import org.folio.circulation.infrastructure.storage.feesandfines.FeeFineRepository;
 import org.folio.circulation.infrastructure.storage.ServicePointRepository;
 import org.folio.circulation.infrastructure.storage.UserRepository;
 import org.folio.circulation.resources.context.RenewalContext;
